+++ src/player/CommentCell.cpp
@@ -285,7 +285,7 @@ class $modify(RLCommentCell, CommentCell) {
             FLAlertLayer::create(
                 "Layout Moderator",
                 "This user can <cj>suggest layout levels</c> for <cl>Rated "
-                "Layouts</c> to the <cr>Layout Admins</c>. They have the ability to <co>moderate the leaderboards</c>.",
+                "Layouts</c> to the <cr>Layout Admins</c>. They have the ability to <co>moderate the leaderboard</c>.",
                 "OK")
                 ->show();
       }
@@ -294,8 +294,8 @@ class $modify(RLCommentCell, CommentCell) {
             FLAlertLayer::create(
                 "Layout Administrator",
                 "This user can <cj>rate layout levels</c> for <cl>Rated "
-                "Layouts</c>. They can change the <cy>featured ranking on the "
-                "featured layout levels.</c>",
+                "Layouts</c>. They have the same power as <cg>Moderators</c> but including the ability to change the <cy>featured ranking on the "
+                "featured layout levels.</c> and <cg>set event layouts</c>.",
                 "OK")
                 ->show();
       }